beats/requirements.txt
flask>=2.0.1,<2.1.0
werkzeug<3.0.0
flask-cors==3.0.10
python-dotenv==0.19.0
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
mutagen==1.45.1
spotipy==2.23.0
requests==2.31.0